feat: add in shortcut call for yaml sequences with single entry
the yaml library in use accepts sequences of arbitrary length, which is totally reasonable, but many
of the entries in yaml files I'm using expect a single entry for a single key. this adds in a
function to directly grab single entries instead of vectors, and the function fails if the requested
key has any number of entries other than one.