fix: attempt to fix std::filesystem crashes
std::filesystem seems to have issues with some types of memory allocations during path construction.
i can't help but wonder if this is a gcc/stl versioning issue, the C++17 support here is really very
close to its first adoption. i'm probably going to fall back to boost::filesystem, as I've
traditionally had very good luck with it, and it's the basis of the std::filesystem spec anyway