fix: address optional trackers and corner case for deleted flags
the tracking_file method check_file was not correctly handling optional flags; this has been
addressed with an additional parameter in its call. additionally that change made clear that a
corner case exists when a tracking file is already present but needs to be removed. also this raises
the issue of default values, which I think has been missed entirely so far, whoopsie