feat: add support for default phenotype config values
in the process of implementing support for missing parameters, I remembered that there's a whole
setup for default values that needed to be added to top level extension config and parsed
appropriately. this patch now adds that support. some flags have defaults and some don't, and config
options that are flagged as essential but lack defaults raise exceptions. the extensible general
extensions are all required to have default values specified, even though they are technically not
essential.