fix: modify tracker output format to comma-delimited lists
one random piece of information I'd forgotten about the existing code is that most multi-token
trackers are emitted as single-line comma-delimited lists, largely for compatibility with command
line calls to other programs. this fix modifies the tracker handler to emit data in that format. the
one exception in the current pipeline is the '.categories' tracker for categorical variable runs
through SAIGE. in this case, i've had the tracking handler always emit comma-delimited lists, but
for categorical variables it's fed a single token which is actually the multiline data that the
current pipeline expects (plus some changes to categorical tracking that need to be implemented in
the model matrix constructor)