fix: correct behavior of categories tracker for binary traits
the *.categories tracker is not being emitted correctly (at all) for binary traits. for categorical
(n>2) traits, it's getting emitted to each subdir comparison*/, but for binary traits it should be
emitted with a "0\treference
1\tcomparison" entry. this patch fixes that. it also adds in the
appropriate test files for `make check`. the *.categories tracker now has its own function, which is
a bit weird, but since it's one of the fixed/required trackers, I guess it's not too much of an
atrocity

re #1