fix: add correct default value for no covariate selection in user config
inherited correct behavior for an analysis model with no covariates was to emit "NA" to the
`*.covariates_selected` tracker, which would be passed to the R model matrix generator. that
behavior was not correctly implemented in v1.0.0; instead an empty tracker file was created. this
patch *should* update that behavior to correctly emit "NA" to such a tracker. test case for this bug
incoming

re #2